refactor: tree-view/model에서 tree-view/utils로 변경
